\name{diggitdata-package}
\Rdversion{1.4}
\alias{diggitdata-package}
\alias{diggitdata}
\docType{package}
\title{
Datasets and network models required for the examples in the diggit package
}
\description{
This package contains a human glioblastoma mRNA expression dataset, a human glioblastoma copy number variation (CNV) dataset, a glioblastoma context-specific transcriptional network, and a glioblastoma context-specific post-translational network.
}
\details{
\tabular{ll}{
Package: \tab diggitdata\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2014-08-29\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Mariano Javier Alvarez

Maintainer: Mariano Javier Alvarez <reef103@gmail.com>
}
