\name{gbm.aracne}
\Rdversion{1.4}
\alias{gbmTFregulon}
\docType{data}
\title{
Transcriptional regulatory network for human glioblatoma
}
\description{
Human glioblastoma transcriptional regulatory network assembled by the ARACNe algorithm from TCGA expression data.
The data is stored in a S3 class 'regulon' object.
}
\usage{data(gbm.aracne)}
\examples{
data(gbm.aracne)
print(gbmTFregulon)
}
\references{
Margolin,A.A. et al. (2006) ARACNE: an algorithm for the reconstruction of gene regulatory networks in a mammalian cellular context. BMC Bioinformatics, 7 Suppl 1, S7.
}
\keyword{datasets}
