\name{gbm.expression}
\Rdversion{1.4}
\alias{gbmExprs}
\docType{data}
\title{
Human glioblastoma mRNA expression dataset from TCGA
}
\description{
ExpressionSet object containing cleaner summarized and mas5 normalized mRNA expression data from 250 human glioblastoma samples profiled by TCGA on HT-HGU133A arrays.
}
\usage{data(gbm.expression)}
\examples{
data(gbm.expression)
print(gbmExprs)
}
\references{
Alvarez,M.J. et al. (2009) Correlating measurements across samples improves accuracy of large-scale expression profile experiments. Genome Biol., 10, R143.
}
\keyword{datasets}
