\name{gbm.mindy}
\Rdversion{1.4}
\alias{gbmMindy}
\docType{data}
\title{
Post-translational regulatory network for human glioblatoma
}
\description{
Human glioblastoma post-translational regulatory network assembled by the MINDy algorithm from TCGA expression data.
The data is stored in a S3 class 'regulon' object.
}
\usage{data(gbm.mindy)}
\examples{
data(gbm.mindy)
print(gbmMindy)
}
\references{
Wang,K. et al. (2009) Genome-wide identification of post-translational modulators of transcription factor activity in human B cells. Nat. Biotechnol., 27, 829-39.
}
\keyword{datasets}
